+++ src/application/Randomizer.h
@@ -39,6 +39,7 @@ public:
     void OnEmblemGoalSet(int emblemGoal);
     void SetStatingArea(StartingArea startingArea);
     void SetMissions(Characters characters, int missions);
+    void SetCharacterLifeSanity(Characters character, bool characterLifeSanity);
     void SetDeathLink(bool deathLinkActive);
     void SetRingLink(bool ringLinkActive);
     void SetHardRingLink(bool hardRingLinkActive);