+++ src/input/archipelago/ArchipelagoManager.cpp
@@ -155,6 +155,36 @@ void SADX_PinballLifeCapsules(const int pinballLifeCapsules)
     randomizerPtr->OnPinballLifeCapsulesSet(pinballLifeCapsules);
 }
 
+void SADX_SonicLifeSanity(const int lifeSanity)
+{
+    randomizerPtr->SetCharacterLifeSanity(Characters_Sonic, lifeSanity);
+}
+
+void SADX_TailsLifeSanity(const int lifeSanity)
+{
+    randomizerPtr->SetCharacterLifeSanity(Characters_Tails, lifeSanity);
+}
+
+void SADX_KnucklesLifeSanity(const int lifeSanity)
+{
+    randomizerPtr->SetCharacterLifeSanity(Characters_Knuckles, lifeSanity);
+}
+
+void SADX_AmyLifeSanity(const int lifeSanity)
+{
+    randomizerPtr->SetCharacterLifeSanity(Characters_Amy, lifeSanity);
+}
+
+void SADX_GammaLifeSanity(const int lifeSanity)
+{
+    randomizerPtr->SetCharacterLifeSanity(Characters_Gamma, lifeSanity);
+}
+
+void SADX_BigLifeSanity(const int lifeSanity)
+{
+    randomizerPtr->SetCharacterLifeSanity(Characters_Big, lifeSanity);
+}
+
 
 void SADX_StartingArea(const int startingArea)
 {
@@ -243,6 +273,13 @@ void ArchipelagoManager::Connect()
     AP_RegisterSlotDataIntCallback("EmblemsForPerfectChaos", &SADX_EmblemsForPerfectChaos);
     AP_RegisterSlotDataIntCallback("LifeSanity", &SADX_LifeSanity);
     AP_RegisterSlotDataIntCallback("PinballLifeCapsules", &SADX_PinballLifeCapsules);
+    AP_RegisterSlotDataIntCallback("SonicLifeSanity", &SADX_SonicLifeSanity);
+    AP_RegisterSlotDataIntCallback("TailsLifeSanity", &SADX_TailsLifeSanity);
+    AP_RegisterSlotDataIntCallback("KnucklesLifeSanity", &SADX_KnucklesLifeSanity);
+    AP_RegisterSlotDataIntCallback("AmyLifeSanity", &SADX_AmyLifeSanity);
+    AP_RegisterSlotDataIntCallback("BigLifeSanity", &SADX_BigLifeSanity);
+    AP_RegisterSlotDataIntCallback("GammaLifeSanity", &SADX_GammaLifeSanity);
+
     AP_RegisterSlotDataIntCallback("StartingArea", &SADX_StartingArea);
     AP_RegisterSlotDataIntCallback("DeathLink", &SADX_SetDeathLink);
     AP_RegisterSlotDataIntCallback("RingLink", &SADX_SetRingLink);