+++ src/output/displayManager/DisplayManager.cpp
@@ -243,7 +243,7 @@ void DisplayManager::DisplayItemsUnlocked()
         const std::string levelInfo = levelName + " (" + characterName + " - " + missionName + ")";
         SetDebugFontColor(currentColor);
         DisplayDebugString(NJM_LOCATION(2, this->_startLine + this->_displayCount+displayOffset), levelInfo.c_str());
-        
+
         // Show missions checked
         displayOffset++;
         buffer.clear();
@@ -255,13 +255,14 @@ void DisplayManager::DisplayItemsUnlocked()
         if (missionsEnabled > 2)
             buffer.append(GetLevelEmblemCollected(&SaveFile, CurrentCharacter, CurrentLevel, MISSION_A) ? " A" : "  ");
         buffer.append(" ");
-        
-        if (CurrentLevel > LevelIDs_HedgehogHammer && CurrentLevel <= LevelIDs_HotShelter)
+
+        if (_options.lifeSanity && _options.GetCharacterLifeSanity(static_cast<Characters>(CurrentCharacter)))
         {
             buffer.append(" Lives: ");
             for (const auto& check : _checkData)
             {
-                if (check.second.character == CurrentCharacter && check.second.type == LocationLifeCapsule && GET_LEVEL(check.second.level) == CurrentLevel )
+                if (check.second.character == CurrentCharacter && check.second.type == LocationLifeCapsule &&
+                    GET_LEVEL(check.second.level) == CurrentLevel)
                 {
                     buffer.append(check.second.checked ? "X" : "-");
                 }
@@ -289,7 +290,6 @@ void DisplayManager::DisplayItemsUnlocked()
         displayOffset++;
     }
 
-   
 
     if (_options.sonicMissions > 0)
     {